social_media_api/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
dj-database-url==3.0.1
django==4.2.25
django-taggit==6.1.0
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==25.0
pillow==10.4.0
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.3
typing-extensions==4.13.2
whitenoise==6.7.0
